Fix 64-bit issues in the SoundCodec plugin.
git-svn-id: http://squeakvm.org/svn/squeak/trunk@3175 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: dc819b10296757236acc5326441a033634afc59b

+++ platforms/Cross/plugins/SoundCodecPrims/SoundCodecPrims.h
@@ -1,17 +1,17 @@
 /* SoundCodec Plugin */
 /* prototypes */
 void gsmEncode(
-	int state, int frameCount,
-	int src, int srcIndex, int srcSize,
-	int dst, int dstIndex, int dstSize,
-	int *srcDelta, int *dstDelta);
+	long state, long frameCount,
+	long src, long srcIndex, long srcSize,
+	long dst, long dstIndex, long dstSize,
+	long *srcDelta, long *dstDelta);
 	
 void gsmDecode(
-	int state, int frameCount,
-	int src, int srcIndex, int srcSize,
-	int dst, int dstIndex, int dstSize,
-	int *srcDelta, int *dstDelta);
+	long state, long frameCount,
+	long src, long srcIndex, long srcSize,
+	long dst, long dstIndex, long dstSize,
+	long *srcDelta, long *dstDelta);
 	
-void gsmInitState(int state);
+void gsmInitState(long state);
 
-int gsmStateBytes(void);
+long gsmStateBytes(void);
